+++ wallet/api/node/mocks/node_mock.go
@@ -36,6 +36,21 @@ func (m *MockGRPCAdapter) EXPECT() *MockGRPCAdapterMockRecorder {
 	return m.recorder
 }
 
+// CheckTransaction mocks base method.
+func (m *MockGRPCAdapter) CheckTransaction(arg0 context.Context, arg1 *v1.CheckTransactionRequest) (*v1.CheckTransactionResponse, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "CheckTransaction", arg0, arg1)
+	ret0, _ := ret[0].(*v1.CheckTransactionResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// CheckTransaction indicates an expected call of CheckTransaction.
+func (mr *MockGRPCAdapterMockRecorder) CheckTransaction(arg0, arg1 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CheckTransaction", reflect.TypeOf((*MockGRPCAdapter)(nil).CheckTransaction), arg0, arg1)
+}
+
 // Host mocks base method.
 func (m *MockGRPCAdapter) Host() string {
 	m.ctrl.T.Helper()